+++ src/application/activity/DeleteActivity.ts
@@ -1,6 +1,6 @@
 import { IActivityRepository } from "../../domain/activity/ActivityRepository";
 import { ILogger } from "../../domain/Logger";
-import { AppError } from "../../errors/errors";
+import { AppError } from "../errors/errors";
 import { HttpStatusCode } from "../../infraestructure/utils/HttpStatusCode";
 
 type DeleteActivityParam = {
@@ -13,18 +13,18 @@ export interface IDeleteActivity {
 export class DeleteActivityService implements IDeleteActivity {
   constructor(
     readonly activityRepository: IActivityRepository,
-    readonly logger: ILogger,
+    readonly logger: ILogger
   ) {}
   async invoke({ id }: DeleteActivityParam): Promise<void> {
     try {
       await this.activityRepository.delete(id);
     } catch (error) {
       this.logger.error(
-        `Some error has been ocurred trying delete a new activity ${error}`,
+        `Some error has been ocurred trying delete a new activity ${error}`
       );
       throw new AppError(
         "Internal server error",
-        HttpStatusCode.InternalServerError,
+        HttpStatusCode.InternalServerError
       );
     }
   }
