+++ src/application/authentication/AuthenticationService.ts
@@ -3,7 +3,7 @@ import { JwtService } from "../../domain/JwtService";
 import { ILogger } from "../../domain/Logger";
 import { Login, LoginService, Token } from "../../domain/LoginService";
 import { IUsersRepository } from "../../domain/users/UsersRepository";
-import { AppError, InvalidCredentials, NotFound } from "../../errors/errors";
+import { AppError, InvalidCredentials, NotFound } from "../errors/errors";
 import { HttpStatusCode } from "../../infraestructure/utils/HttpStatusCode";
 
 export class AuthenticationService implements LoginService {
@@ -11,7 +11,7 @@ export class AuthenticationService implements LoginService {
     readonly usersRepository: IUsersRepository,
     readonly jwtService: JwtService,
     readonly bcrypt: HashService,
-    readonly logger: ILogger,
+    readonly logger: ILogger
   ) {}
   async invoke({ email, password }: Login): Promise<Token> {
     const existentUser = await this.usersRepository.findByEmail(email);
@@ -23,14 +23,14 @@ export class AuthenticationService implements LoginService {
 
     const isValidPassword = await this.bcrypt.compare(
       password,
-      existentUser.password,
+      existentUser.password
     );
 
     if (!isValidPassword) {
       this.logger.warn(`User credential are invalid ${email}`);
       throw new InvalidCredentials(
         "Invalid credentials",
-        HttpStatusCode.Unauthorized,
+        HttpStatusCode.Unauthorized
       );
     }
 
@@ -39,11 +39,11 @@ export class AuthenticationService implements LoginService {
       return { token };
     } catch (error) {
       this.logger.error(
-        `Some internal server error has been ocurred trying log user : ${error}`,
+        `Some internal server error has been ocurred trying log user : ${error}`
       );
       throw new AppError(
         "Internal server error",
-        HttpStatusCode.InternalServerError,
+        HttpStatusCode.InternalServerError
       );
     }
   }
