+++ src/infraestructure/express/controllers/activity/GetActivityController.ts
@@ -1,7 +1,7 @@
 import { Request } from "express";
 import { HttpStatusCode } from "../../../utils/HttpStatusCode";
 import { z } from "zod";
-import { Controller, HttpResponse } from "../../../../../domain/Controller";
+import { Controller, HttpResponse } from "../../../../domain/Controller";
 import { IGetActivity } from "../../../../application/activity/GetActivityService";
 
 export const zodValidationActivitySchema = z.object({
