+++ src/infraestructure/express/controllers/activity/ListActivityController.ts
@@ -1,7 +1,7 @@
 import { Request } from "express";
 import { HttpStatusCode } from "../../../utils/HttpStatusCode";
 import { z } from "zod";
-import { Controller, HttpResponse } from "../../../../../domain/Controller";
+import { Controller, HttpResponse } from "../../../../domain/Controller";
 import { IListActivity } from "../../../../application/activity/ListActivityService";
 
 export const zodValidationActivitySchema = z.object({
