+++ src/infraestructure/logger/LoggerService.ts
@@ -13,7 +13,7 @@ class LoggerService implements ILogger {
         format.colorize(),
         format.printf(({ level, message, timestamp }) => {
           return `${timestamp} ${level}: ${message}`;
-        })
+        }),
       ),
       transports: [
         new transports.Console(),
